vertexai/model_deployment/requirements.txt
sentence-transformers
numpy
google-cloud-aiplatform
torch
fastapi[standard]
google-cloud-aiplatform[prediction]==1.16.0